requirements.txt
langchain
langchain-google-genai
langchain-chroma
langgraph
pydantic
python-dotenv